style: Move display outside of mako
