+++ components/style/values/computed/box.rs
@@ -9,8 +9,8 @@ use values::computed::length::LengthOrPercentage;
 use values::generics::box_::AnimationIterationCount as GenericAnimationIterationCount;
 use values::generics::box_::VerticalAlign as GenericVerticalAlign;
 
-pub use values::specified::box_::{AnimationName, OverflowClipBox, OverscrollBehavior};
-pub use values::specified::box_::{ScrollSnapType, TouchAction, WillChange};
+pub use values::specified::box_::{AnimationName, Display, OverflowClipBox};
+pub use values::specified::box_::{OverscrollBehavior, ScrollSnapType, TouchAction, WillChange};
 
 /// A computed value for the `vertical-align` property.
 pub type VerticalAlign = GenericVerticalAlign<LengthOrPercentage>;
